Create lily_value_structs.h from lily_core_types.h
lily_core_types.h is now focused entirely on structs holding
information for the frontend of the interpreter, moreso than what vm
or ts or backend parts need.

The new lily_value_structs.h is intended to be used internally only.
The intent is to provide a lily_api_value.h which contains high-level
value operations. lily_value_structs.h will thus be included when a
part of the interpreter needs to reach into the plumbing.